+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * `LimitStatus` is `Comparable<LimitStatus>` since this update
 * `Limits` is `Comparable<Limits>` since this update
+* Main API has been changed
 
 ## 0.13.0